Add files via upload
My Portfolio Website
Welcome to my portfolio website! This is a personal project where I've showcased my skills, experience, and some of the work I have done. The website is built using HTML and CSS to demonstrate my understanding of web development fundamentals.

Features:
A responsive layout to ensure a great user experience across all devices.
A clean and minimalist design to highlight the content and my work.
Sections including an introduction, portfolio, and contact information.
Technologies Used:
HTML: For creating the structure of the website.
CSS: For styling the layout and making it visually appealing.
Feel free to explore the code, and don't hesitate to provide feedback or suggestions. Thank you for checking out my portfolio!